Refactor unit tests to remove mockResizeObserver and implement ResizeObserver in Jest setup
