+++ front/app/components/admin/Graphs/MultiBarChart/MultiBarChart.test.tsx
@@ -4,7 +4,6 @@ import { NilOrError } from 'utils/helperUtils';
 import { render, screen, waitFor } from 'utils/testUtils/rtl';
 
 import { legacyColors } from '../styling';
-import { mockResizeObserver } from '../testSetup';
 
 import MultiBarChart from './';
 
@@ -14,10 +13,6 @@ const getEmptyData = (): Row[] | NilOrError => null;
 const getErrorData = (): Row[] | NilOrError => new Error();
 const emptyData = getEmptyData();
 const errorData = getErrorData();
-
-beforeAll(() => {
-  mockResizeObserver();
-});
 
 describe('<MultiBarChart />', () => {
   describe('Missing data', () => {
