+++ front/app/components/admin/Graphs/PieChart/PieChart.test.tsx
@@ -3,8 +3,6 @@ import React from 'react';
 import { NilOrError } from 'utils/helperUtils';
 import { render, screen, waitFor } from 'utils/testUtils/rtl';
 
-import { mockResizeObserver } from '../testSetup';
-
 import PieChart from './';
 
 type Row = { a: number; name: string };
@@ -17,10 +15,6 @@ const data: Row[] = [
 const getNilData = (): Row[] | NilOrError => null;
 const getErrorData = (): Row[] | NilOrError => new Error();
 const getData = (): Row[] | NilOrError => data;
-
-beforeAll(() => {
-  mockResizeObserver();
-});
 
 describe('<PieChart />', () => {
   describe('Missing data', () => {
