+++ front/internals/jest/setup.js
@@ -4,7 +4,107 @@ import 'whatwg-fetch';
 
 Object.assign(global, { TextDecoder, TextEncoder });
 
-global.ResizeObserver = require('resize-observer-polyfill');
+// Mock ResizeObserver to work in JSDOM (which has no layout engine).
+// Parses width/height from element styles/attributes.
+function parsePixelValue(value) {
+  if (!value || typeof value !== 'string') return NaN;
+  if (value.endsWith('px')) return parseInt(value, 10);
+  if (value.endsWith('%')) return NaN;
+  const num = parseInt(value, 10);
+  return value === String(num) ? num : NaN;
+}
+
+function getMockDimensions(element) {
+  let width = parsePixelValue(
+    element?.getAttribute('style')?.match(/width:\s*(\d+px)/)?.[1]
+  );
+  let height = parsePixelValue(
+    element?.getAttribute('style')?.match(/height:\s*(\d+px)/)?.[1]
+  );
+
+  if (!(width > 0)) width = parsePixelValue(element?.getAttribute('width'));
+  if (!(height > 0)) height = parsePixelValue(element?.getAttribute('height'));
+
+  let parent = element?.parentElement;
+  while (parent && (!(width > 0) || !(height > 0))) {
+    if (!(width > 0)) {
+      width =
+        parsePixelValue(
+          parent.getAttribute('style')?.match(/width:\s*(\d+px)/)?.[1]
+        ) || parsePixelValue(parent.getAttribute('width'));
+    }
+    if (!(height > 0)) {
+      height =
+        parsePixelValue(
+          parent.getAttribute('style')?.match(/height:\s*(\d+px)/)?.[1]
+        ) || parsePixelValue(parent.getAttribute('height'));
+    }
+    parent = parent.parentElement;
+  }
+
+  return { width: width > 0 ? width : 400, height: height > 0 ? height : 200 };
+}
+
+class MockResizeObserver {
+  constructor(callback) {
+    this.callback = callback;
+  }
+
+  observe(target) {
+    const { width, height } = getMockDimensions(target);
+    queueMicrotask(() => {
+      this.callback(
+        [
+          {
+            target,
+            contentRect: {
+              width,
+              height,
+              top: 0,
+              left: 0,
+              bottom: height,
+              right: width,
+              x: 0,
+              y: 0,
+              toJSON: () => ({}),
+            },
+            borderBoxSize: [{ inlineSize: width, blockSize: height }],
+            contentBoxSize: [{ inlineSize: width, blockSize: height }],
+            devicePixelContentBoxSize: [
+              { inlineSize: width, blockSize: height },
+            ],
+          },
+        ],
+        this
+      );
+    });
+  }
+
+  unobserve() {}
+  disconnect() {}
+}
+
+global.ResizeObserver = MockResizeObserver;
+
+// Mock getBoundingClientRect for consistent dimensions in JSDOM
+const originalGetBoundingClientRect = Element.prototype.getBoundingClientRect;
+Element.prototype.getBoundingClientRect = function () {
+  const { width, height } = getMockDimensions(this);
+  if (width > 0 || height > 0) {
+    return {
+      width,
+      height,
+      top: 0,
+      left: 0,
+      bottom: height,
+      right: width,
+      x: 0,
+      y: 0,
+      toJSON: () => ({}),
+    };
+  }
+  return originalGetBoundingClientRect.call(this);
+};
 
 HTMLCanvasElement.prototype.getContext = jest.fn(() => ({
   globalAlpha: 1,
